hgmo: add more Mercurial 4.6 compatibility changes (Bug 1469673) r=gps
After upgrading the Mercurial version running in the hgmo virtual
environments themselves, several missed API changes were detected.
The value returned from Mercurial's `wireproto.clonebundles` is
now a `bytesresponse` object, and the same data must be parsed
from the `data` attribute. `mercurial.wireproto` no longer exists,
and is therefore wrapped in an `import_module` call where necessary.
Parsing of request headers has moved to a different attribute, and
the need to append "HTTP" to the front of the desired header,
as well as the requirement to change any header "-" to "_" is
gone.

Differential Revision: https://phabricator.services.mozilla.com/D2529

--HG--
extra : moz-landing-system : lando