ansible: add optional arguments to skip instance types during deploy r=gps
Currently when deploying to hg.mozilla.org, there is a single
deploy process that deploys the current code to all web instances
in the cluster. In some cases we only need to deploy to hgweb
or hgssh instances to send out an update to a particular
application. This commit adds the `--skip-{hgssh,hgweb}` flags,
which will force the deploy process to skip making any changes
to hgssh or hgweb hosts.

Differential Revision: https://phabricator.services.mozilla.com/D2568

--HG--
extra : moz-landing-system : lando