testing: upgrade Mercurial test harness
Upgraded to version is based on the files in the Mercurial 4.7
release (33ac6a72308a).

We preserved changes for:

* Increased port count
* Path for finding hghave
* DOCKER_HOSTNAME
* Setting ERR_FILE

We also dropped the removal of various [defaults] entries related to
default `-d` arguments for setting dates because they are needed to
test legacy Mercurial versions.

We also removed the updating of the TESTDIR environment variable. It
now (properly) refers to the directory of the individual test. But
most of our tests are relying on it pointing to the repo root. So
we'll need a follow-up to fix this.