Bug 1480232 - robustcheckout: only use scmutil.revsingle; r=sheehan
repo.__contains__ behaves differently in 4.7+ and only allows specific
values. scmutil.revsingle() is the preferred lookup for user-defined
values that we don't want to interpret as revsets.

The code was previously using revsingle() but only after a
repo.__contains__ invocation.

We change the code to always use revsingle().

This makes all the developer warnings in the tests go away.

Differential Revision: https://phabricator.services.mozilla.com/D2633

--HG--
extra : moz-landing-system : lando