vcsreplicator: get path to Mercurial executable from config file during bootstrap r=gps
The current implementation requires passing the path to the replication
virtualenv on the command line for the hgweb process. However the
vcsreplicator daemons already use the config file to determine where
the replication virtualenv is located. This commit changes the bootstrap
process to resolve the vcsreplicator virtualenv from the config file.

Differential Revision: https://phabricator.services.mozilla.com/D2670

--HG--
extra : moz-landing-system : lando