unifyrepo: abort when source commit is not shared by all repos (Bug 1484231) r=gps
If this warning is displayed, we will always hit an abort
later in the program: "abort: repository is unrelated". We will hit
this condition while pulling in changesets from other repos, but
we may end up wasting time pulling in changesets for repos
that are relevant first. We should abort upon hitting this
condition instead.

Differential Revision: https://phabricator.services.mozilla.com/D3933

--HG--
extra : moz-landing-system : lando