unifyrepo: set `path` variable correctly while iterating over source repos (Bug 1484231) r=gps
The path variable is set when iterating over the source repos
the first time, but it is never reset to the correct
source when iterating over the sources a second time. Therefore
whatever repo was last in the iteration during the first
loop will have the same name in the final loop.

This causes an issue in the `unifypushes` function, as we will
hit the `if lastsource == source and lastid == pushid` condition
when pushlogs between two separate repos have the same push id.
For example, this condition will be true for repo1 push 1 and
repo2 push 1, as source is the same for all repos. This is likely
the root cause of bug 1480203, further testing will be required
to know for certain.

Differential Revision: https://phabricator.services.mozilla.com/D3935

--HG--
extra : moz-landing-system : lando