unifyrepo: move `hg replicatesync` call behind conditional flag (Bug 1484231) r=gps
Without this flag, `unifyrepo` is dependent on vcsreplicator
to be tested. This means we require the hgmodocker environment
to run tests on unifyrepo, and we currently don't have those
features available in CI. With this flag we can turn off
the replication call and test the extension without Docker.

Differential Revision: https://phabricator.services.mozilla.com/D3934

--HG--
extra : moz-landing-system : lando