unifyrepo: set `testedwith` to 4.6, acquire wlock before bookmark write
Now that we have tests for this extension, we can set the
`testedwith` string as we do with other extensions. Since we
only need the extension to work under 4.6 at the moment,
this commit also gets rid of some compatibility conditionals.

After adding this change, the tests give a devel-warn about
requiring a `wlock()` as well as a `lock()` to write bookmarks.
This commit also adds that to the extension.

--HG--
extra : amend_source : ad86ebccd8e6f2669d6e25141aa39c690d07e3bf