docs: update references to `mach mercurial-setup` (Bug 1488101)
The version-control-tools docs reference the `mach mercurial-setup`
command in quite a few places. This command was changed to
`mach vcs-setup` in bug 1257478 without updating the documentation
to reflect the change. This commit globally replaces `mercurial-setup`
with `vcs-setup`.

Additionally, some of the documentation for legacy extensions (such as
bzexport and the mq extensions) reference the setup command despite
having been removed as an install step some time ago. This commit
also updates the documentation to indicate the extension must be
installed manually.