pushlog: send `pushdate` through `hgdate` filter (Bug 1488405) r=gps
The raw output of the `pushdate` template keyword is a tuple
of (timestamp, tzoffset). When the template layer displays
this tuple, it simply concatenates the two values together and
returns the result as a string. In reality, we only want the
`timestamp` portion of this tuple. On the hgweb servers, the
tzoffset value is always `0`, which is causing the timestamps
emitted in pulse/SNS notifications from the server to appear
to be off by a factor of 10. The `hgdate` template filter
takes a date tuple and formats it with the string "%d %d".

This commit sends the output of the `pushdate` template keyword
through the `hgdate` template filter during the call to `hg log`
which returns the information used in push notifications (pulse
and SNS).

Differential Revision: https://phabricator.services.mozilla.com/D5469

--HG--
extra : moz-landing-system : lando