unifyrepo: move stage repo setup immediately before repo unification (Bug 1480203) r=gps
Currently the unification process creates the staging
repo at the start of the command, does intermediate
processing to determine which changesets/pushlog
entries will need to be pulled in, then performs
the repo unification. During the intermediate
step, some new pushes may enter the source repos and
are collected as resources to add to the unified
repository - however these new pushes do not exist
in the staging repo and the process fails to complete
(most annoyingly, failing to update bookmarks).

This commit moves the staging repo pull step closer
to the end of the process. With this change, the process
should determine which actions it needs to take then
create the stage repo. This will ensure that no changeset
present in the set of required unification actions is
missing from the stage repo.

Differential Revision: https://phabricator.services.mozilla.com/D5808

--HG--
extra : moz-landing-system : lando