vcsreplicator: move `pushdataaggregator_groups` from /etc/mercurial to /repo/hg (Bug 1491149) r=gps
This commit changes the canonical path of the "pushdataaggregator_groups"
file from /etc/mercurial to /repo/hg. With this change, the
file is now on the shared NFS mount and is accessible by the
ssh failover server.

I have already copied the file to the new location, so this should
"just work" once we deploy the change. Then we will just need to
delete the old file.

Differential Revision: https://phabricator.services.mozilla.com/D5809

--HG--
extra : moz-landing-system : lando