pushlog: rename `id` variables to `pushid` (Bug 1485878) r=gps
`id` is the name of a built-in function in Python. Creating a
variable of this name overrides the existing `id` function and
is considered a bad practice. We already reference this field as
`pushid` in other places (including the core pushlog extension),
so this commit changes the feed extension to do the same.

Differential Revision: https://phabricator.services.mozilla.com/D6298

--HG--
extra : moz-landing-system : lando