pushlog: create entries in the `pushlog_feed` webcommand with a generator (Bug 1485878) r=gps
This commit changes the `entries` field in the template mapping
to be generated using a `mappinggenerator`. Setting the entries
field to a list as in the current implementation breaks under
Mercurial 4.7, so this change makes us compatible in both versions.

In the new implementation, we use a generator expression instead
of a for loop to iterate through the query entries. To determine
whether a changeset is hidden or not, we use the `isrevsymbol`
function instead of catching a `FilteredRepoLookupError`. We
also use `revsingle` to get the context variable instead of
using `repo[node]`.

In making this change, I noticed that the `test-obsolescence-pushlog.t`
test started to fail. Upon further review, I realized that the test
is expecting the obsolete changesets in the test repo to not appear
in the ATOM output, but in the previous implementation they were
showing up anyways. This is because in the previous implementation
we would check for a hidden changeset, but then `pass` instead of
`continue` in the except block. I have removed the bad output and
the behaviour now meets the specification in the test.

Differential Revision: https://phabricator.services.mozilla.com/D6304

--HG--
extra : moz-landing-system : lando