pushlog: implement pushlog's `changelist` as a generator (Bug 1485878) r=gps
This commit makes a variety of changes to the `changelist` function
defined within the `pushloghtml` webcommand.

First, the function is no longer a closure defined within the
`pushloghtml` webcommand, instead becoming a function at the
top-level of the module. Next, the limit variable is changed to
a `tiponly` flag, as the  value was only ever set to 0 (no limit)
or 1 (single entry returned).

This commit then makes changes to the way `changelist` yields entries
to the template layer. Formerly, we would process every entry in
the pushlog query and aggregate them in a list, then use a slice of
that list and yield entries. Thus, we were processing every push in
a given query and then only returning a single entry (in the worst
case). This commit instead makes the generator yield entries as they
are ready, avoiding the extra unnecessary work.

In the previous implementation, we would iterate over every push in
the query and create a template mapping for that entry. Along the
way, we would make several logical decisions that influence which
fields are populated and how. The first entry for each pushid has
a "push" field with the user and time of the push. If the push was
a merge, we also add a "mergerollup" field which displays the number
of changesets added by this merge to the first entry for that push.
To populate the number, we would create a reference to the "push"
mapping for that entry and increment by 1 for each successive entry
which had the same pushid.

The new implementation separates the logic into easier to understand
chunks. `pushlog_changelist` now contains the logic to iterate over
different pushes and make decisions about how to display each set
of entries for a push. We iterate over all the entries and aggregate
pushes into a queue until the current pushid changes, at which point
we use `handle_entries_for_push` to yield entries with the correct
information.

`handle_entries_for_push` yields entries given the gathered context
for a single push. The "mergerollup" count field is populated as the
length of the queue of entries for that push. We also determine whether
a push is a merge and whether to hide the merge fields in this section
of code.

`create_entry` takes the contexts gathered in `handle_entries_for_push`
and returns a template mapping for that entry, which is then yielded
successively down the line to the template layer.

Differential Revision: https://phabricator.services.mozilla.com/D6308

--HG--
extra : moz-landing-system : lando