ansible: grab hg-deploy-bot IRC password from the server (Bug 1491116)
Right now our Ansible deploys are broken due to a +R setting on
the #vcs IRC channel that prevents the playbook from sending
a notification due to lack of registration. I have registered
a password for the hg-deploy-bot and placed it in a file on
the server. Hopefully with this workaround we can get through
a deployment without a manual IRC notification.