ansible: remove hgssh1 as hard-coded master ssh server r=gps
I noticed while fixing bug 1491116 that the host we slurp the
mirror keys and bot password from is set to the hgssh-prod-master
Ansible group, but we always access the variable from hgssh1.
If we happen to failover to the backup server and need to deploy
something, this is going to slow us down. Might as well fix it now.

Differential Revision: https://phabricator.services.mozilla.com/D6401

--HG--
extra : moz-landing-system : lando