zookeeper: mirror systemd config in test suite (Bug 1492453); r=gps
In a previous commit we added GC logging to the Zookeeper processes
by setting environment variables in the systemd unit for those
processes. It slipped my mind that we don't use systemd in
docker, and need to update in a second location for parity in
the test suite. This commit should fix that issue.

Differential Revision: https://phabricator.services.mozilla.com/D6447

--HG--
extra : histedit_source : c72fd79f8ef75660e59da4224c893cd220ad952f