hgmo: add `automationrelevance` JSON template (Bug 1485874) r=gps
This commit adds an `automationrelevance` template to the `json`
template, allowing us to send the output of the `automationrelevance`
webcommand through the template layer instead of streaming
JSON directly. Due to this change, output from the raw
HTTP endpoint is no longer nicely indented. We now send the
response to a body file and then use `json.tool` to format
the output.

Differential Revision: https://phabricator.services.mozilla.com/D6736

--HG--
extra : moz-landing-system : lando