vcsreplicator: add logic to parse `replicationrules` config section (Bug 1491373) r=gps
This commit adds logic to the `config.Config` constructor for
parsing a config section with rules for filtering repositories
out of the set of repositories to replicate. The section must
be named `replicationrules` and each entry takes the form
`behaviour.name = syntax:rule`. "Behaviour" is one of
{include, exclude}, indicating if the rule will force inclusion
or exclusion of any matching repos. "Name" is an alias given
to the rule, which will be used for debugging and logging
during filtering (something like "repo X filtered by rule Y").
"Syntax" is the type of rule the "rule" string will be applied
to. The current available "syntax" values are "re" for regular
expression and "path" for explicit (ie explicitly include repo X).
An example section would look like:

    [replicationrules]
    exclude.user = re:\{moz\}/user.*
    include.releases = re:\{moz\}/releases/.*
    include.myuserrepo = path:{moz}/users/cosheehan_mozilla.com/vct
    exclude.beta = path:{moz}/releases/mozilla-beta

Follow up commits will specify the ordering of rule checking and
how the replication processes will use these filters.

Differential Revision: https://phabricator.services.mozilla.com/D6913

--HG--
extra : moz-landing-system : lando