vcsreplicator: create a `repofilter` decorator and apply to message handlers (Bug 1491373) r=gps
This commit creates a repository filtering decorator which wraps
the message handler function to check if a message corresponds
to a repo which is being filtered out on this instance. A
decorator is used to separate the filtering logic out from the
message handler logic, while allowing the filter logic to be defined
in a single place.

To test this functionality, we create a new `filteredconsumer` which
has a set of replication rules defined in it's configuration file.
We also add some output to `test-corrupt-repo-replication.t` to
account for the extra layer on the failure stack trace created by
the `repofilter` wrapper.

Differential Revision: https://phabricator.services.mozilla.com/D6917

--HG--
extra : moz-landing-system : lando