vcsreplicator: add a `filter` function to the `Config` class (Bug 1491373) r=gps
This commit adds a function which applies filter rules to a given
repo name (assumed to be in wire path format) and returns a
`RepoFilterResult` with a yes/no boolean indicating if the
repo should be filtered or not, and a string name value
which is the alias of the rule which did/did not filter
the repo. The rules are applied in the following order

    0. If there are no filters defined, pass.
    1. If repo is explicitly included, pass.
    2. If repo is explicitly excluded, fail.
    3. If repo matches an include regex and
       does not match an exclude regex, pass.
    4. Catch-all fail.

Using this order will allow us to set general pattern
matching rules to include/exclude repos, but also bypass
those rules for specific repositories as we deem necessary.

Differential Revision: https://phabricator.services.mozilla.com/D6914

--HG--
extra : moz-landing-system : lando