vcsreplicator: apply repo filtering to hgweb bootstrap procedure (Bug 1491373) r=gps
This commit makes the bootstrap procedure respect the
repo filtering rules set in the hgweb replicator config
file. The hgweb procedure receives the input object
from the hgssh procedure and applies the aggregate filter
function to the set of repos specified in the hgweb config
file. Then the process continues as normal, except when a
message is being processed for a repo which has been removed
from the set of repos to replicate, it is ignored and the
process continues. Immediately after removing filtered repos
from the set of repos to be replicated, we also log which
repos were filtered out and specify the offending rule for
debugging purposes.

There are a few advantages to this implementation. The hgssh
step is unaware of the filtering rules applied to the individual
hgweb instances. This means we could set up separate instances
to be dedicated to different repositories in the future and
bootstrap all of them with the result of a single hgssh bootstrap
run. We also keep the hgssh and hgweb roles separate in terms of
shared data since we won't need to pass the hgweb replication config
to the hgssh instances for filtering.

A disadvantage to this approach is that we will spend some time
collecting and acknowledging Kafka messages for repos we wish to
replicate. If we did pass hgweb filter rules to the hgssh instance,
we could avoid aggregating the messages altogether and only produce
bootstrap messages for repos we wish to replicate. This would have
a disadvantage in being need to run multiple times if we wish
to bootstrap multiple hgweb heads with different replication rules
at the same time.

To test this feature, we add a `set-config-option` file to the
Ansible docker-hgweb role, which is essentially the same as
docker-hgssh's `set-hgrc-option`, except it takes a full path
to the file and acts on any valid `.ini` file.

Differential Revision: https://phabricator.services.mozilla.com/D6916

--HG--
extra : moz-landing-system : lando