pushlog: enable compatibility with Mercurial 4.8 (Bug 1501686) r=gps
This commit makes pushlog tests pass under Mercurial 4.8rc0.
The only code change required is calling `report` as a private
method under Mercurial 4.8, as not doing so causes the rollback
to fail without explanation why. The other changes are related
to differing test output in the new version, notably a fixed bug that
caused phase data to not be transferred during a stream clone
has been changed - the test for that behaviour in
`test-pull-obsolete.t` is put under a 4.8 conditional.

Differential Revision: https://phabricator.services.mozilla.com/D10067

--HG--
extra : moz-landing-system : lando