serverlog: mark as compatible with Mercurial 4.8 (Bug 1501686) r=gps
This commit removes the Mercurial 4.6 output from the
serverlog tests and adds output from version 4.8. There
is still a `TRACKING` flag related to the potential
unification of SSH and HTTP log events - since the
extension works under 4.8 that code is left as-is
until we find time to implement that change.

Differential Revision: https://phabricator.services.mozilla.com/D10108

--HG--
extra : moz-landing-system : lando