hgmo: add no-op arguments to `allparents` call under 4.8 (Bug 1503308) r=gps
In the previous commit we removed the 4.6 version of this
function, and in this commit we add the 4.8 version. The
new signature takes `context` and `mapping` to align
with other templatekeywords, however neither are used
as the function will take `ctx` from the closure instead.
We pass `None` as both arguments and function executes as
expected.

Differential Revision: https://phabricator.services.mozilla.com/D10253

--HG--
extra : moz-landing-system : lando