hgmo: convert repo path to byte string (Bug 1503308) r=gps
Under Mercurial 4.8, the repo path passed to the
`hg.repository` constructor must be string-like
instead of unicode (the format the data is in while
in the `data` dict). Converting to `bytes` aligns
us with the conversions in this function and makes
the hgmo extension compatible with Mercurial 4.8.

Differential Revision: https://phabricator.services.mozilla.com/D10254

--HG--
extra : moz-landing-system : lando