pash: add `scm_versioncontrol` to ssh group membership check r=gps
:jlund came into #vcs this week and requested a way to
test if someone has push access to a repo. :kwierso
mentioned the `ssh hg.mozilla.org` tool, but noted that
it did not mention his ability to push to version control
tools. This is likely due to the pash utility being fairly
old and predating the scm_versioncontrol group. This
commit adds scm_versioncontrol to the mapping of groups
to repositories, so people with access to version control
tools can verify using pash.

Differential Revision: https://phabricator.services.mozilla.com/D9913

--HG--
extra : moz-landing-system : lando