hgserver: update tests for 4.8 compatibility (Bug 1501686) r=gps
The tests under `hgserver/` pass, with the exception of some
extra output under 4.8. There is also a test which checks for
`mozreview-ldap-associate` which is EOL and failing under
4.7 as well, this commit removes that section.

`test-bookmarks-replication.t` is still intermittently
failing due to a bookmark not showing up as replicated
on the first call to the `json-bookmarks` webcommand -
subsequent calls show it exists so I suspect there is
a race condition there.

Differential Revision: https://phabricator.services.mozilla.com/D10390

--HG--
extra : moz-landing-system : lando