docs: add setuptools as a dependency; r=sheehan
blockdiag has a dependency on setuptools, which pip-compile ignores
by default. We need to list it so its hashes are included in order
to appease `pip install --require-hashes`, which we use.

This change makes `create-environment docs` work. Unfortunately,
docs building in the new virtualenv still encounters a number of
errors...

Differential Revision: https://phabricator.services.mozilla.com/D10401

--HG--
extra : moz-landing-system : lando