docs: use and upgrade latest versions of packages; r=sheehan
Now that we use pip-compile and have a separate virtualenv for docs
that won't interfere with the testing virtualenv, we don't need to
worry about conflicting package versions. We are free to use the
latest Sphinx packages for building the docs. So let's do that.

Because we check in the requirements file with pinned versions,
things should still be deterministic. We only need to check compat
when manually upgrading packages.

Differential Revision: https://phabricator.services.mozilla.com/D10403

--HG--
extra : moz-landing-system : lando