env: make test-requirements.txt managed by pip-compile; r=sheehan
pip-compile can follow .in files to pull in dependencies. So we
point it at deploy-requirements.in instead of deploy-requirements.txt.

pip-compile simply emits a complete manifest of all packages. So the
`-r deploy-requirements.txt` from the now-generated
test-requirements.txt is dropped and packages from
deploy-requirements.in are defined inline. This is fine.

Differential Revision: https://phabricator.services.mozilla.com/D10468

--HG--
extra : moz-landing-system : lando