ansible/hg-web: install mod_wsgi manually; r=sheehan
pip-compile has problems with the mod_wsgi package. I'm guessing its
packaging is really wonky since it is an Apache module and not a
Python package per se.

This commit removes mod_wsgi from the requirements file and puts it
under the domain of Ansible and a manual pip package install.

Differential Revision: https://phabricator.services.mozilla.com/D10600

--HG--
extra : moz-landing-system : lando