ansible: upgrade to Ansible 2.0; r=sheehan
We are running Ansible 1.9, which is relatively old and has known
security vulnerabilities. I initially held off upgrading because Ansible
2 was much slower. Enough time has passed that we are too far out of
date on Ansible and we can no longer live with the security issues.
Let's upgrade to a newer Ansible.

Ansible 2.7 is the latest Ansible. But there are a number of
backwards incompatible changes to it that make the jump straight to
it pretty difficult. Our approach to the upgrade will be to make
smaller jumps so we can understand them easier.

Our first step is Ansible 2.0.

The porting guide to 2.0 can be found at
https://docs.ansible.com/ansible/latest/porting_guides/porting_guide_2.0.html.

While it lists a number of changes, I couldn't find any obvious
violations in our Ansible code! However, I did not verify every
single Ansible playbook. So there may be legacy Ansible code lingering.
This caveat applies to all future patches as well.

Differential Revision: https://phabricator.services.mozilla.com/D10657

--HG--
extra : moz-landing-system : lando