firefoxtree: enable compatibility with Mercurial 4.8 (Bug 1501686) r=gps
In versions above 4.8, the old-style template keyword
definitions will be deprecated. This commit forces
the `fxheads` template keyword to be dynamically
defined with the new API in versions above 4.6, and
with the old API for versions below 4.6. Doing so
fixes a deprecation warning in tests on version
4.8 and makes us compatible with modern hg.

Differential Revision: https://phabricator.services.mozilla.com/D10917

--HG--
extra : moz-landing-system : lando