hgweb: use fastannotate from core Mercurial (Bug 1504794) r=gps
`fastannotate` is now shipped with core Mercurial. We no longer
need to pull down and build the hg-experimental repo as part
of our deploy process. This commit purges hg-experimental from
vct (with the exception of a few extensions recommended in the
docs).

We also flip the `server = True` flag to serve the annotate cache
over the wire protocol.

Differential Revision: https://phabricator.services.mozilla.com/D11124

--HG--
extra : moz-landing-system : lando