robustcheckout: record more granular metrics for overall times (bug 1506958); r=sheehan
Currently, we measure the times of individual operations as well as an
"overall" time. A problem with this approach is that it is difficult
to make sense of what different "flavors" of operations are doing. e.g.
in order to differentiate the "overall" time for a fresh clone from
that of an incremental update, you need to look at all other reported
metrics for a sample and see if certain actions are reported. This
requires complexity on the analysis side and requires the analysis
code to make assumptions about how the metrics collection works. This
is complicated and brittle because assumptions may change.

This commit introduces several new "overall_*" metrics that report
the "overall" time N ways depending on various factors. This should
allow us to better track e.g. overall times for fresh clones versus
overall times for pull+update. This should make it much easier to
conduct data analysis.

Another way of capturing this data in the Perfherder format is by
using "extraOptions." I'm tempted to shoehorn the set of collected
"behaviors" into "extraOptions." But this commit doesn't do that
just yet.

The impetus for this change is wanting to have better metrics around
partial clones so we can measure the impact they have on CI.

Differential Revision: https://phabricator.services.mozilla.com/D11808

--HG--
extra : moz-landing-system : lando