extensions: introduce `wireprotocache` extension; r=gps
Mercurial 4.8 has been released and will soon be deployed
to hg.mo. Once that happens, we will want to make changes
to the web heads to leverage the new wire protocol and all
it's crazy features. One of the key features of this new
protocol is the ability to generically cache command responses
by monkeypatching several key functions. Another is the ability
to issue a content redirect response to clients, telling them
the data they need can be found at some other locations.

This commit introduces the `wireprotocache` extension.
This extension provides a set of drop-in plugins (currently
limited to S3) which implement cachers of various types for
wire protocol responses.

The only implemented plugin thus far is an S3 plugin.
Servers add their AWS access key and key ID to an hgrc config,
and specify the name of the S3 bucket which holds the objects.
When a cache lookup request comes in, the cacher sends a HEAD
request to S3 which will return a 404 if the object does not
exist (ie a cache miss). If the request is a cache hit, a presigned
url for the object is generated and used to issue a content
redirect response which is sent to the client. If the response
indicates a cache miss, the response is generated by the server
and buffered in the cache until `onfinished` is called. During
`onfinished`, we calculate the size of the response and can
optionally avoid caching if the response is below a configured
minimum threshold. Otherwise we insert the object into the
cache bucket using the `put_object` API.

To test the S3 plugin, we require the `moto` mock AWS library.
Specifically, we use the "standalone server" functionality,
which creates a Flask application that imitates S3. A new hghave
predicate is added to check for this functionality before
testing.

Should we choose to switch cache providers in the future,
(for example, GCP's blob store, Redis, memcached,
etc) we will simply need to write the plugin and drop it into
the extension directory.

Differential Revision: https://phabricator.services.mozilla.com/D11109

--HG--
rename : hgext/unifyrepo/tests/hghave => hgext/wireprotocache/tests/hghave
extra : amend_source : e26e9de10960746ae61dddc8f30d754df7e509dd