hgmo: wire protocol command to expose raw cache files (bug 1507036); r=sheehan
Early experiments with partial clone show us burning a few dozen
seconds of CPU during clone computing the branchmap cache. This
is because wireprotov2 doesn't yet transfer the branchmap cache
natively.

Because a few dozen seconds in this context is a significant
percentage of execution time, it's worth our time to improve matters.

This commit implements a new wire protocol command that exposes the
raw data of known cache files in the .hg/cache directory of the
repository. Clients can call this command to essentially
copy the raw cache files.

All the cache files are optional and clients should be able to
handle malformed or corrupt files. So it should be harmless to
transfer all available cache files.

Differential Revision: https://phabricator.services.mozilla.com/D11965

--HG--
extra : moz-landing-system : lando