terraform: create terraform directory with `privsubnet` module (Bug 1509039) r=gps
This commit creates a top-level `terraform` directory which will
serve as the root module (with root module files to be added in
a later commit) for all vcs-related Terraform configs going
forward. Within this directory we add a `modules` subdirectory
and populate it with the `privsubnet` module. This module
configures a private subnet within a specified VPC, region
and availability zone. The subnet creates a route table, which
causes all outbound traffic from instances within the subnet to
be routed to a specified NAT gateway (in practice this will be
the NAT gateway of a public subnet in the same AZ).

As this is the first addition of the "terraform" directory,
and Terraform itself creates local state files, we add an entry
to `.hgignore` to avoid tracking unnecessary files.

Differential Revision: https://phabricator.services.mozilla.com/D12915

--HG--
extra : moz-landing-system : lando