terraform: add `pubsubnet` module (Bug 1509039) r=gps
This commit adds a `pubsubnet` module to the Terraform
configs. The module configures a public subnet within a
specified VPC, and attaches a NAT gateway that can receive
traffic from a private subnet (from the same AZ in practice).
A route table is taken as input, which should include a
rule that sends outbound traffic within the subnet (including
traffic sent from a private subnet through the NAT gateway)
to the internet gateway attached to the VPC.

Differential Revision: https://phabricator.services.mozilla.com/D12916

--HG--
extra : moz-landing-system : lando