terraform: add an `hgweb-mirror` module (Bug 1509039) r=gps
This commit adds an `hgweb-mirror` module to the Terraform
configs. This module creates a single EC2 instance intended
to be configured as an hgweb mirror. The new instance is
created using the CentOS7 AMI and are intended to be assigned
to a private subnet within a VPC. The instances use the c4.2xl
instance type and are assigned to security groups via the
`security_group_ids` variable.

Differential Revision: https://phabricator.services.mozilla.com/D12918

--HG--
extra : moz-landing-system : lando