terraform: add the `ci-only` module (Bug 1509039) r=gps
This commit adds the `ci-only` module, which applies
the config for a VPC containing all CI-only resources
in a single AWS region. The module creates a VPC
and attaches an internet gateway with an assigned
elastic IP address. A route table is created with
a rule to send all public subnet traffic to the
internet gateway. We then create a public and private
subnet for each AZ withing the region, assigning a
/19 CIDR block to each. A bastion host is created
within a public subnet in the VPC to facilitate SSH
access to mirrors in the VPC (for both administrative
access and Ansible deployments). Finally we create a
security group and network ACls for the hosts
and assign relevant security rules to each.

Differential Revision: https://phabricator.services.mozilla.com/D12919

--HG--
extra : moz-landing-system : lando