terraform: add root module files (Bug 1509039) r=gps
This commit adds the files for the Terraform root module.
When running Terraform commands from a directory, any `.tf`
files in that directory are considered to be within the
`root` module, and all other infrastructure should be
contained within the root module. We add the necessary
module files and configure a single `ci-only` module,
assigned to the `us-west-2` AWS region. As we expand
our infrastructure to new AWS regions or new cloud
providers (GCP, etc), we will simply add new module
entries to the root module (for example add a new
`ci-only` module entry for the us-east-1 region).

Differential Revision: https://phabricator.services.mozilla.com/D12920

--HG--
extra : moz-landing-system : lando