robustcheckout: remove phase check; r=sheehan
Per the inline comment, this check was added when we had a replication lag
window resulting in phases not being transferred properly. The server-side issue
is fixed and we no longer need this sanity check.

The experimental partial clone support may not transfer phases data properly
and was triggering this check. But that's expected. Only a few tasks care
about phases and we can opt those out of partial clone mode.

Differential Revision: https://phabricator.services.mozilla.com/D13909

--HG--
extra : moz-landing-system : lando