terraform: add s3 module for various S3 buckets (Bug 1510795) r=gps
This commit adds an S3 module which configures buckets for
various hg projects within different regions. As of now,
we only create the clone/pull bundle packages in this module.
In the near future we will add IAM policies to this module that
allow storing wire protocol responses as well.

We also create an associated logging bucket with the
`log-delivery-write` ACL to log bucket events.

Differential Revision: https://phabricator.services.mozilla.com/D13856

--HG--
extra : moz-landing-system : lando