testing: install `Pygments` in test suite to restore colored output
Last month we performed a large-scale upgrade of various packaging
dependencies for version-control-tools work. In that upgrade, we
removed the "Pygments" package, which was being used to provide
color output to version-control-tools test results. This commit
re-adds Pygments as a dependency to the test suite and is confirmed
to restore colored test output.

--HG--
extra : amend_source : 789a6090c0fc925a8209be10350079021923ad27